Proof-of-concept for HHS/mesh-rdf/161 and HHS/mesh-rdf/156:
Lodestar makes limited use of Spring - using it only to:
 - handle creation and management of singletons such as controllers and SparqlServices
 - provide back-end URL mappings for calls that get back-end resources

To fix 156 and 161, we need to do a SPARQL lookup in a controller, pass the result through a model, and
then render it in a view.   Add to this that we already have the controller "ExplorerServlet"
and the view "explore.jsp"

This fix adds a proof-of-concept that we can use more Spring webmvc to get this done using the
existing controller and view.